# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"AppResponseStatusEnum": {
				"enum": [
					"success",
					"error",
					"TOKEN_UPDATED",
					"TOKEN_EXPIRED",
					"OTP_SENT"
				],
				"type": "string"
			},
			"AppResponse__username-string--email-string__": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/AppResponseStatusEnum"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"email": {
								"type": "string"
							},
							"username": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"username"
						],
						"type": "object"
					}
				},
				"required": [
					"statusCode",
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUserDTO": {
				"properties": {
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"fullName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AppResponse__status-string__": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/AppResponseStatusEnum"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"status": {
								"type": "string"
							}
						},
						"required": [
							"status"
						],
						"type": "object"
					}
				},
				"required": [
					"statusCode",
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AppResponse__token-any--refreshToken-any__": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/AppResponseStatusEnum"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"refreshToken": {},
							"token": {}
						},
						"required": [
							"refreshToken",
							"token"
						],
						"type": "object"
					}
				},
				"required": [
					"statusCode",
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AppResponse__username-string--email-string__"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterUserDTO"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/AppResponse__status-string__"
										},
										{
											"$ref": "#/components/schemas/AppResponse__token-any--refreshToken-any__"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/verify": {
			"post": {
				"operationId": "VerifyAccount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AppResponse__token-any--refreshToken-any__"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"otp": {
										"type": "number",
										"format": "double"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"otp",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8000/api/v1",
			"description": "Development server"
		},
		{
			"url": "https://allnepalibusiness/api/v1",
			"description": "production server"
		}
	]
}